% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_age_hdi.R
\name{age_samples}
\alias{age_samples}
\title{Generate sample for age estimate}
\usage{
age_samples(age_est, n)
}
\arguments{
\item{age_est}{Result of call to estimate_dental_age()}

\item{n}{Number of samples}
}
\value{
n samples
}
\description{
Generate sample for age estimate
}
