% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DentalAge.R
\docType{package}
\name{DentalAge-package}
\alias{DentalAge-package}
\alias{DentalAge}
\title{DentalAge}
\description{
Estimate dental age from scores of dental development.
}
\details{
\tabular{ll}{ Package: \tab DentalAge\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2021-06-25\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{package}
