% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_age_hdi.R
\name{estimate_age_hdi}
\alias{estimate_age_hdi}
\title{Estimate HDI from dental age estimate}
\usage{
estimate_age_hdi(age_sd, n = 1e+05, interval = 0.5)
}
\arguments{
\item{age_sd}{numeric vector output from \code{estimate_dental_age()}.}

\item{n}{numeric number of samples to produce. Defaults to 1e5 for speed.
Increase this number for investigating the tails of the distribution.}

\item{interval}{numeric width of the HD interval. Defaults to 0.5}
}
\value{
numeric vector with the upper and lower bounds of the interval
}
\description{
Estimate HDI from dental age estimate
}
\examples{
means <- get_means_for_scores(x = ExampleScores[1, ])
age <- estimate_dental_age(means)
estimate_age_hdi(age)
}
