% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dental_age.R
\name{estimate_dental_age}
\alias{estimate_dental_age}
\title{Estimate dental age}
\usage{
estimate_dental_age(means)
}
\arguments{
\item{means}{matrix (6 x 2) of means and standard deviations}
}
\value{
numeric vector with estimated age and variance
}
\description{
Given a matrix of means and sds, estimate dental age following the method
decribed in Šešelj M, Sherwood RJ, Konigsberg LW. 2019. Timing of
Development of the Permanent Mandibular Dentition: New Reference Values
from the Fels Longitudinal Study. Anat Rec 302:1733–1753.
}
\examples{
means <- get_means_for_scores(x = ExampleScores[1, ])
estimate_dental_age(means)

}
