% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleScores}
\alias{ExampleScores}
\title{Example dental scores}
\format{
A data frame with 5 rows and 8 variables:
\describe{
\item{ID}{individual ID}
\item{Sex}{character of sex: "F" or "M"}
\item{Canine}{character score for canine}
\item{P3}{character score for P3}
\item{P4}{character score for P4}
\item{M1}{character score for M1}
\item{M2}{character score for M2}
\item{M3}{character score for M3}
}
}
\usage{
ExampleScores
}
\description{
A dataset containing five example dental scores for use in calculating
dental age.
}
\keyword{datasets}
