% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_means_for_scores.R
\name{get_means_for_scores}
\alias{get_means_for_scores}
\title{Lookup mean and sd for a sex / tooth / stage combination}
\usage{
get_means_for_scores(x, verbose = TRUE)
}
\arguments{
\item{x}{data frame with one row}

\item{verbose}{boolean flag for printing diagnostic messages}
}
\value{
6 x 2 matrix with means and standard deviations for each tooth.
}
\description{
This function expects a row with columns for "Canine", "P3", "P4", "M1",
"M2", and "M3" that contain scores (or NA). A matrix of means and standard
deviations for each tooth is returned.
}
\examples{
get_means_for_scores(x = ExampleScores[1, ])

# Example from Seselj et al. (2019)
x <- data.frame(Sex = "M", Canine = "R.25", P3 = "R.25",
                P4 = "R.i", M1 = "A.5",
                M2 = "R.25", M3 = NA)
means <- get_means_for_scores(x)
estimate_dental_age(means)

}
