% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_score.R
\name{recode_score}
\alias{recode_score}
\title{Recode scores}
\usage{
recode_score(x)
}
\arguments{
\item{x}{string stage coding to correct}
}
\value{
string of recoded stage score
}
\description{
Recode scores
}
\examples{
recode_score("zero")
recode_score("Ci")

}
