% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_to_stage.R
\name{score_to_stage}
\alias{score_to_stage}
\title{Convert stage numeric score to character score}
\usage{
score_to_stage(x)
}
\arguments{
\item{x}{vector of numeric stage scores}
}
\value{
vector with numeric scores converted to character
}
\description{
Convert stage numeric score to character score
}
\examples{
data.frame(numeric_stage = 1:14, stage = score_to_stage(c(1:14)))

}
