% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_score.R
\name{validate_score}
\alias{validate_score}
\title{Validate stage score}
\usage{
validate_score(x)
}
\arguments{
\item{x}{string stage score to check}
}
\value{
Boolean of the score's validity
}
\description{
Validate stage score
}
