% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_means_for_scores.R
\name{get_means_for_scores}
\alias{get_means_for_scores}
\title{Lookup mean and sd for a sex / tooth / stage combination}
\usage{
get_means_for_scores(x)
}
\arguments{
\item{x}{Row data frame}
}
\value{
6 x 2 matrix with means and standard deviations for each tooth.
}
\description{
This function expects a row with columns for "Canine", "P3", "P4", "M1",
"M2", and "M3" that contain scores (or NA). A matrix of means and standard
deviations for each tooth is returned.
}
\examples{
get_means_for_scores(x = ExampleScores[1, ])

}
